--- src/main/java/com/ureca/juksoon/domain/feed/entity/FileType.java
@@ -1,8 +1,0 @@
-package com.ureca.juksoon.domain.feed.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum FileType {
-    IMAGE, VIDEO
-}
